Allow strings as names of the request parameters
Background:
Phoenix handles request parameters as map with string keys.
Phoenix_params take these parameters from the conn, validates them
according to the definition provided by the user and puts them back
into the conn.
In the request definition only atoms were allowed as names of the
request parameters. So, when put back into the conn, the user should
have started working with atoms instead of strings.

In some cases it is more convinient for the user to continue using
strings instead of atoms. So, phoenix_params allow request parameters to
be define as string and when put back into the conn, the user will
continue working with strings instead of atoms.

The string representation of the parameters should remain unique among
all paramters. It means - it does not matter whether the parameter name
is string or atom, their string representations should not be equal.

Eg. It is now allowed to have `:token` and `"token"` as different
parameters because their string representations are the same.